requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==6.1.0
Markdown==3.4.1
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.15.0
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==6.1.0
Markdown==3.4.1
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.15.0
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==6.1.0
Markdown==3.4.1
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
zipp==3.15.0
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==6.1.0
Markdown==3.4.1
psycopg-binary==3.1.8
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
zipp==3.15.0
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==6.1.0
Markdown==3.4.1
psycopg-binary==3.1.8
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
zipp==3.15.0
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.1.0
Markdown==3.4.1
psycopg-binary==3.1.8
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
zipp==3.15.0
asgiref==3.6.0
Cerberus==1.3.4
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.1.0
Markdown==3.4.1
psycopg-binary==3.1.8
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
serpy==0.3.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
zipp==3.15.0
